Renaming `ElementConfigurationContainer` to `ElementConfigurationSectionNew` (tmp name).
